Switched the UIManager to be a regaular class, accessed via GameManager
This makes a lot more sense in the new architecture. Everything will be
accessed through the GameManager, and whoever calls the GameManager is
responsible for setting the UI property to be the class it wants to manage
the UI Elements. If it is not set, then the default will be to use the CLI
UI.