Changed the way Victory Points are handled in the Player class.
Previously, when a player built a building, the VP from that building were
immediately added to the Player's VP property. This has been changed. The
previous VP property has been renamed to BoardVP and buildings are now
included when a new property, TotalVP is called.

This more closely mimics the way the physical board game tracks VP, and
makes it much clearer for debugging. It will also be necesarry to have
these properties split apart in the XNA version so that the board can
be drawn with the right number of VP displayed on it.